battery-status/battery-discharging-05-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Status Test: BatteryManager - onlevelchange (discharging)</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/battery-status/">
<meta name="flags" content="interact">
<meta name="assert" content="Decrease the level attribute\'s value and fire a levelchange event when the battery level is updated">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <h2>Preconditions</h2>
  <ol>
    <li>
      The device is unplugged from the charger before this test is run.
    </li>
    <li>
      The battery must neither be empty or full, nor reach empty or full capacity during the test.
    </li>
  </ol>
</body>
<script>

setup({ explicit_timeout: true });

async_test(function (t) {
  navigator.getBattery().then(
    function (battery) {
      var battery_level = battery.level;
      battery.onlevelchange = t.step_func(function (e) {
        assert_true(e.level < battery_level, 'The value of the level attribute must decrease.');
        t.done();
      });
    },
    function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  )
}, document.title);

</script>
